pricing pixed
