+++ src/pages/Home/index.js
@@ -1,5 +1,6 @@
 export {default as Home} from './Home';
-export {default as About} from './About';
-export {default as RequestAppointment} from './RequestAppointment';
-export {default as OurServices} from './OurServices';
 export {default as Hero} from './Hero';
+export {default as About} from './About';
+export {default as OurServices} from './OurServices';
+export {default as RequestAppointment} from './RequestAppointment';
+export {default as TestimonialSection} from './TestimonialSection/TestimonialSection';
